requirements.txt
beautifulsoup4==4.11.1
pandas==1.2.4
python_dateutil==2.8.2
requests==2.25.1
